moved server user only classes from common module into server
