--- commons/src/main/scala/it/parttimeteam/entities/Player.scala
@@ -1,3 +1,0 @@
-package it.parttimeteam.entities
-
-trait Player extends Identifiable with Referable
