+++ server/src/main/scala/it/parttimeteam/common/GamePlayer.scala
@@ -1,4 +1,4 @@
-package it.parttimeteam.entities
+package it.parttimeteam.common
 
 import akka.actor.ActorRef
