+++ server/src/main/scala/it/parttimeteam/common/Identifiable.scala
@@ -1,4 +1,4 @@
-package it.parttimeteam.entities
+package it.parttimeteam.common
 
 trait Identifiable {
