+++ server/src/main/scala/it/parttimeteam/common/Player.scala
@@ -1,0 +1,3 @@
+package it.parttimeteam.common
+
+trait Player extends Identifiable with Referable
