+++ server/src/test/scala/it/parttimeteam/lobby/LobbySpec.scala
@@ -1,7 +1,7 @@
 package it.parttimeteam.lobby
 
+import it.parttimeteam.common.{GamePlayer, Player}
 import it.parttimeteam.lobby.{GameLobby, Lobby}
-import it.parttimeteam.entities.{GamePlayer, Player}
 import org.scalamock.scalatest.MockFactory
 import org.scalatest.wordspec.AnyWordSpecLike
